# Overlay handling with fuse-overlayfs