This implements #928: Use [fuse-overlayfs](https://github.com/containers/fuse-overlayfs/) as fallback for the kernel overlayfs, with the goal of providing a solution for #776 and #1067.

Discussion of progress is in #1036.