@PhilippWendler 2024-07-09
If I try this out on Ubuntu 22.04 with `bin/containerexec bash`, it just hangs after mounting (last line of debug log is `Mounting '/run' as hidden`). I have to kill the processes with `kill -9`. Any idea what is going on?