@PhilippWendler 2024-09-17
> Thank you for your kind words! I've been celebrating the Mid-Autumn Festival recently, so I didn't have the chance to reply to the comments earlier.

No worries.

> I have submitted two new patches to address the two points that needed modification from your comments.

Thanks!
I think now we just need to find a fix for non-overlay on `/tmp` and then it is done.

> > `benchexec test/tasks/benchmark.xml` is noticeably slower than `benchexec --read-only-dir / test/tasks/benchmark.xml` for me. Is this the same for you?
> 
> I don't think there is a noticeable difference in execution speed between the two on my machine. I conducted the tests on Ubuntu 24.04 with fuse-overlayfs 1.13. I've used `time` command to measure.

Hm, but you also have a difference from 1.2s to 1.8s. For me it was like 1.2s to 2.2s, so same order of magnitude. But I think we can accept this for now and see how it works in practice. Optimizations are fine to add later. Thanks for testing!